# api/openapi.yaml
openapi: 3.0.3
info:
  title: Roller_hoops API
  version: 0.1.0
  description: |
    Canonical API contract for the Roller_hoops system.

servers:
  - url: /api

tags:
  - name: Devices
  - name: Discovery

paths:
  /v1/devices:
    get:
      tags: [Devices]
      summary: List devices
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
    post:
      tags: [Devices]
      summary: Create a device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/devices/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Devices]
      summary: Get device by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Devices]
      summary: Update device by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/discovery/run:
    post:
      tags: [Discovery]
      summary: Trigger a discovery pass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryRunRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryRun'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/discovery/status:
    get:
      tags: [Discovery]
      summary: Get discovery status
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryStatus'

components:
  schemas:
    Device:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          description: Optional human-friendly name.
        metadata:
          $ref: '#/components/schemas/DeviceMetadata'
    DeviceCreate:
      type: object
      description: |
        Minimal create payload. Additional fields will be added as the data model is finalized.
      properties:
        display_name:
          type: string
        metadata:
          $ref: '#/components/schemas/DeviceMetadata'
    DeviceUpdate:
      type: object
      description: |
        Minimal update payload. Additional fields will be added as the data model is finalized.
      properties:
        display_name:
          type: string
        metadata:
          $ref: '#/components/schemas/DeviceMetadata'
    DeviceMetadata:
      type: object
      properties:
        owner:
          type: string
        location:
          type: string
        notes:
          type: string
    DiscoveryRunRequest:
      type: object
      properties:
        scope:
          type: string
          description: Optional scope hint for the discovery engine.
    DiscoveryRun:
      type: object
      required: [id, status, started_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        scope:
          type: string
          nullable: true
        stats:
          type: object
          additionalProperties: true
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
    DiscoveryStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: |
            High level status: "idle" when no runs exist, otherwise mirrors the latest run status.
        latest_run:
          $ref: '#/components/schemas/DiscoveryRun'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true